core-api/requirements.txt
flake8==3.7.9
flake8-docstrings==1.5.0
black==19.10b0
pytest==5.3.4
flask==1.1.1
flask-graphql==2.0.1
gunicorn==20.0.4
graphene==2.1.8
snapshottest==0.5.1
graphene-sqlalchemy==2.2.2
psycopg2==2.8.4
